split ht_color examples and add ht compass example
